backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-jazzmin==3.0.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
Pillow==10.2.0
whitenoise==6.6.0
